+++ src/parser.c
@@ -23,17 +23,17 @@ static StartPositionArray start_position_array_ctor(size_t capacity);
 static int start_position_array_dtor(StartPositionArray* array);
 
 // TODO make full version with many commands support
-int parse_commands(int argc, char* argv[], CommandsArray* commands_array) {
+int parse_commands(size_t argc, char* argv[], CommandsArray* commands_array) {
     assert(argv);
     assert(commands_array);
     
     if (argc < 1) return 0;
 
-    StartPositionArray command_start_positions = start_position_array_ctor((size_t) argc);
+    StartPositionArray command_start_positions = start_position_array_ctor(argc);
 
     command_start_positions.array[command_start_positions.current_pointer].start = 0;  
     
-    for (size_t i = 0; i < (size_t) argc; i++) {
+    for (size_t i = 0; i < argc; i++) {
         if (!strcmp(argv[i], AND_COMMAND_TOKEN)) {
             command_start_positions.array[command_start_positions.current_pointer].end = i - 1;
